moved block testutil to testutil package
